+++ src/examples/source/BenchmarkPixelsGpu.cu
@@ -26,10 +26,12 @@ int main(int argc, char** argv)
 	parseCmdArgs(argc, argv, data_dir, n_dims, data_size, in_memory, shard_span_perc, n_engines);
 
 	// Setup the Benchmarking suite
-	Benchmark1TimePointGpu<int, float> benchmark_1_tp;
+	//Benchmark1TimePointGpu<int, float> benchmark_1_tp;
+  Benchmark1TimePointGpu<int, int> benchmark_1_tp;
 
 	// Setup the TensorCollectionGenerator
-	TensorCollectionGeneratorGpu<int, float> tensor_collection_generator;
+	//TensorCollectionGeneratorGpu<int, float> tensor_collection_generator;
+  TensorCollectionGeneratorGpu<int, int> tensor_collection_generator;
 
 	// Setup the device
   cudaStream_t stream;
